Extract SpeechToTextConfig from configs
